modularize the remaining legacy ingame interface widgets
and consolidate chrome root initialization
