+++ mods/ra/rules/world.yaml
@@ -6,7 +6,6 @@ World:
 	ScreenMap:
 	ActorMap:
 	LoadWidgetAtGameStart:
-		Widget: INGAME_ROOT
 	ScreenShaker:
 	MenuPaletteEffect:
 	WaterPaletteRotation:
