+++ mods/ts/maps/blank-shellmap/map.yaml
@@ -41,7 +41,6 @@ Rules:
 		-SpawnMPUnits:
 		-MPStartLocations:
 		LoadWidgetAtGameStart:
-			Widget: MAINMENU
 
 Sequences:
